requirements.txt
amqp==5.1.1
aniso8601==9.0.1
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Cython==0.29.32
dnspython==2.2.1
email-validator==1.3.0
ffprobe==0.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
gTTS==2.3.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
minio==7.1.12
mongoengine==0.24.2
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.36
pydub==0.25.1
PyJWT==2.6.0
pymongo==4.3.3
PyMuPDF==1.21.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
tesseract==0.1.3
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
